feat(runtime-shared): Notification port + Electron/Web adapters (ADR 0006)
Introduce `NotificationPort` at `packages/runtime-shared/src/ports/notification.ts`
with two abstract methods: `sendNotification(payload)` and `beep()`.

`ElectronNotification` delegates to `new Notification(…).show()` and `shell.beep()`;
`WebNotification` uses the browser Notification API (permission-guard preserved) and
logs a no-op warning for beep.  Both adapters are wired into the `Providers` interface
and their respective `Runtime` instantiations.  Each IPC notification-service handler
shrinks to a single delegating line.
