+++ packages/runtime-shared/src/base.ts
@@ -5,6 +5,7 @@ import type { Logger } from 'tslog';
 
 import type CredentialsProvider from './ports/credentials';
 import type AesProvider from './ports/encryption-aes';
+import type NotificationPort from './ports/notification';
 import type PreferencesStore from './ports/preferences-store';
 import type ProjectOpener from './ports/project-opener';
 import type StorageProvider from './ports/storage';
@@ -56,6 +57,7 @@ export interface Providers {
 	aes: AesProvider;
 	credentials: CredentialsProvider;
 	logger: Logger<unknown>;
+	notification: NotificationPort;
 	preferences: PreferencesStore;
 	projectOpener: ProjectOpener;
 	storage: StorageProvider<GenericStore>;
